rm indirection for express package
